+++ src/components/Text/Text.tsx
@@ -36,7 +36,7 @@ const Text = <As extends keyof JSX.IntrinsicElements>(props: T.Props<As>) => {
 		color && s[`--color-${color}`],
 		...responsiveClassNames(s, "--variant", variant),
 		...responsiveClassNames(s, "--align", align),
-		weight && s[`--weight-${weight}`],
+		...responsiveClassNames(s, "--weight", weight),
 		decoration && s[`--decoration-${decoration}`],
 		maxLines !== undefined && s[`--clamp`],
 		maxLines === 1 && s["--break-all"],